funcion agregada *Promedio semestral* ya aparece en VentanaSemestre el promedio semestral

+++ src/promedianainterfaz/Materia.java
@@ -25,7 +25,7 @@ public class Materia {
         this.nombre = nombre;
     }
 
-    public double getCreditos() {
+    public int getCreditos() {
         return creditos;
     }
